Module Std.Applicable

module type S = sig ... end

Applicable is an abstraction lying between Arrow, Monad and Applicative. It can be seen as a more general form of Applicative that lacks return (or pure) operation. Since return operation is not available for co-inductive type, we cannot use the Applicative functor for Streams.