Module Std.Signal

A handler to produce elements in streams.

type 'a t = 'a signal

val send : 'a signal -> 'a -> unit

send s x sends value x to an associated stream.

val repeat : 'a signal -> times:int -> 'a -> unit

repeat s ~times:n x sends value x to an associated stream n times