Module type Applicable.S

Applicable is an abstraction lying between Arrow, Monad and Applicative. It can be seen as a more general form of Applicative that lacks return (or pure) operation. Since return operation is not available for co-inductive type, we cannot use the Applicative functor for Streams.

type 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

map xs ~f transform xs with f

val apply : ('a -> 'b) t -> 'a t -> 'b t

apply fs xs apply functions fs to xs